connect program to categories table in db

+++ src/main/java/entities/Category.java
@@ -7,9 +7,9 @@ import java.util.Random;
 public class Category {
     private final String title;
     private final String description;
-    private final Double id;
+    private final long id;
     Random rand = new Random();
-    public Category( Double id,String title, String description) {
+    public Category( long id,String title, String description) {
         this.title = title;
         this.description = description;
         this.id = id;
